+++ learning/katas/go/core_transforms/map/lesson-info.yaml
@@ -19,5 +19,5 @@
 
 content:
 - pardo
-- ParDo OneToMany
+- pardo_onetomany
 - ParDo Struct
