+++ learning/katas/go/core_transforms/map/pardo_onetomany/pkg/task/task.go
