+++ learning/katas/go/core_transforms/map/pardo_onetomany/task-info.yaml
@@ -32,7 +32,3 @@ files:
     placeholder_text: TODO()
 - name: cmd/main.go
   visible: true
-- name: go.mod
-  visible: false
-- name: go.sum
-  visible: false
