+++ learning/katas/go/core_transforms/map/pardo_onetomany/task-remote-info.yaml
