+++ learning/katas/go/core_transforms/map/pardo_onetomany/test/task_test.go
@@ -13,29 +13,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package test
 
 import (
-	"context"
+	"beam.apache.org/learning/katas/core_transforms/map/pardo_onetomany/pkg/task"
 	"github.com/apache/beam/sdks/go/pkg/beam"
-	"github.com/apache/beam/sdks/go/pkg/beam/log"
-	"github.com/apache/beam/sdks/go/pkg/beam/x/beamx"
-	"github.com/apache/beam/sdks/go/pkg/beam/x/debug"
-	"pardo_onetomany/pkg/task"
+	"github.com/apache/beam/sdks/go/pkg/beam/testing/passert"
+	"github.com/apache/beam/sdks/go/pkg/beam/testing/ptest"
+	"testing"
 )
 
-func main() {
+func TestTask(t *testing.T) {
 	p, s := beam.NewPipelineWithRoot()
-
-	input := beam.Create(s, "Hello Beam", "It is awesome")
-
-	output := task.ApplyTransform(s, input)
-
-	debug.Print(s, output)
-
-	err := beamx.Run(context.Background(), p)
-
-	if err != nil {
-		log.Exitf(context.Background(), "Failed to execute job: %v", err)
+	tests := []struct {
+		input beam.PCollection
+		want []interface{}
+	}{
+		{
+			input: beam.Create(s, "Hello Beam. It is awesome."),
+			want: []interface{}{"Hello", "Beam.", "It", "is", "awesome."},
+		},
+	}
+	for _, tt := range tests {
+		got := task.ApplyTransform(s, tt.input)
+		passert.Equals(s, got, tt.want...)
+		if err := ptest.Run(p); err != nil {
+			t.Error(err)
+		}
 	}
 }
